caboca

+++ utils.c
@@ -3,8 +3,8 @@
 void get_gauss(float mean, float stddev, float *u, float *v) {
     float x, y, s = 0.0f;
     while (s >= 1.0f || s == 0.0f) {
-        x = (rand() / ((float)RAND_MAX)) * 2.0f - 1.0f;
-        y = (rand() / ((float)RAND_MAX)) * 2.0f - 1.0f;
+        x = ((float)rand() / ((float)RAND_MAX)) * 2.0f - 1.0f;
+        y = ((float)rand() / ((float)RAND_MAX)) * 2.0f - 1.0f;
         s = x * x + y * y;
     }
     s = sqrtf(-2.0f * logf(s) / s);
